CHANGED: major refactor if ModPackPanes
 * Moved similar code from FTBPacksPane and ThirdPartyPane into
   new AbstractModPackPane.java
 * Converted private/static variables and functions mostly publics
 * Renamed function
 * Fixed other classes to call member function instead of calling static
   funtions
